Minor changers

+++ src/main/java/com/example/java/maven/ticketVendingMachine/Ticket.java
@@ -17,7 +17,7 @@ public enum Ticket {
         this.symbol = symbol;
     }
 
-    public float getValue() {
+    public int getValue() {
         return value;
     }
